[FEAT] rebuild guias form
